Move legacy code to  module
